# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Management Platform",
...
    "termsOfService": "https://api-management-platform.com/terms",
    "contact": {
      "name": "API Support Team",
      "url": "https://api-management-platform.com/support",
      "email": "support@api-management-platform.com"
    },
    "license": {
      "name": "Proprietary License",
      "url": "https://api-management-platform.com/license"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List Users",
        "description": "List all users (Admin only)\n\nReturns paginated list of users with filtering and sorting.\n\n**Query Parameters:**\n- page: Page number (default: 1)\n- limit: Items per page (1-100, default: 20)\n- search: Search by email, username, or full name\n- role: Filter by role (user, admin, super_admin)\n- status: Filter by status (active, inactive, suspended, deleted)\n- sort_by: Sort field (created_at, email, username, last_login_at)\n- sort_order: Sort order (asc, desc)\n\n**Headers:**\n- Authorization: Bearer {admin_access_token}\n\n**Response:**\n- items: List of users\n- total: Total count\n- page: Current page\n- limit: Items per page\n- pages: Total pages",
        "operationId": "list_users_api_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by email, username, or name",
              "title": "Search"
            },
            "description": "Search by email, username, or name"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by role",
              "title": "Role"
            },
            "description": "Filter by role"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort order (asc/desc)",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order (asc/desc)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_UserListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/api-keys": {
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key\n\nCreates a new API key for the authenticated user. The full API key is\nreturned only once - save it securely!\n\n**Request Body:**\n- name: Descriptive name for the key (required)\n- description: Optional description\n- organization_id: Optional organization ID (for organization keys)\n- scopes: List of permission scopes (default: [\"read:*\"])\n- ip_whitelist: List of allowed IP addresses (optional)\n- allowed_origins: List of allowed origins for CORS (optional)\n- rate_limit_per_hour: Custom hourly rate limit (optional)\n- rate_limit_per_day: Custom daily rate limit (optional)\n- expires_in_days: Expiration in days (optional, no expiration if not set)\n\n**Headers:**\n- Authorization: Bearer {access_token}\n\n**Response:**\n- id: API key ID\n- key: **Full API key (only shown once!)**\n- name, description, scopes, etc.\n\n**IMPORTANT:** The full API key is only returned once. Store it securely!",
        "operationId": "create_api_key_api_v1_api_keys_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyWithSecret"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List My Api Keys",
        "description": "List API keys\n\nReturns paginated list of API keys for the authenticated user.\n\n**Query Parameters:**\n- page: Page number (default: 1)\n- limit: Items per page (1-100, default: 20)\n- organization_id: Filter by organization (optional)\n\n**Headers:**\n- Authorization: Bearer {access_token}\n\n**Response:**\n- items: List of API keys (without full key!)\n- total: Total count\n- page: Current page\n- limit: Items per page\n- pages: Total pages\n\n**Note:** Full API keys are never returned in list endpoints, only\nthe prefix and last 4 characters.",
        "operationId": "list_my_api_keys_api_v1_api_keys_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by organization",
              "title": "Organization Id"
            },
            "description": "Filter by organization"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ApiKeyListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webhooks": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Create Webhook",
        "description": "Create a new webhook\n\nCreates a webhook endpoint to receive event notifications.\n\n**Request Body:**\n- url: Webhook endpoint URL (must be HTTPS in production)\n- events: List of events to subscribe to\n  - Available events: api_key.created, api_key.deleted, payment.succeeded,\n    payment.failed, subscription.created, subscription.cancelled, etc.\n- description: Optional description\n- organization_id: Optional organization ID\n- is_active: Enable/disable webhook (default: true)\n- secret: Optional webhook secret for signature verification (auto-generated if not provided)\n\n**Headers:**\n- Authorization: Bearer {access_token}\n\n**Response:**\n- Webhook configuration with generated secret\n\n**Important:** Save the webhook secret securely. Use it to verify webhook\nsignatures to ensure requests are coming from our platform.",
        "operationId": "create_webhook_api_v1_webhooks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "List Webhooks",
        "description": "List webhooks\n\nReturns paginated list of webhooks for the authenticated user.\n\n**Query Parameters:**\n- page: Page number (default: 1)\n- limit: Items per page (1-100, default: 20)\n- organization_id: Filter by organization (optional)\n- is_active: Filter by active status (optional)\n\n**Headers:**\n- Authorization: Bearer {access_token}\n\n**Response:**\n- items: List of webhooks\n- total: Total count\n- page: Current page\n- limit: Items per page\n- pages: Total pages",
        "operationId": "list_webhooks_api_v1_webhooks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by organization",
              "title": "Organization Id"
            },
            "description": "Filter by organization"
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by active status",
              "title": "Is Active"
            },
            "description": "Filter by active status"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WebhookListResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "ip_whitelist": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ip Whitelist"
          },
          "allowed_origins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Origins"
          },
          "rate_limit_per_hour": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Hour"
          },
          "rate_limit_per_day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Limit Per Day"
          },
          "expires_in_days": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 365.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires In Days"
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ApiKeyCreate",
        "description": "Schema for creating an API key"
      },
...
      "ApiKeyWithSecret": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "last_four": {
            "type": "string",
            "title": "Last Four"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count"
          },
          "rate_limit_per_hour": {
            "type": "integer",
            "title": "Rate Limit Per Hour"
          },
          "rate_limit_per_day": {
            "type": "integer",
            "title": "Rate Limit Per Day"
          },
          "key": {
            "type": "string",
            "title": "Key"
          }
        },
        "type": "object",
        "required": [
          "created_at",
          "updated_at",
          "name",
          "id",
          "key_prefix",
          "last_four",
          "scopes",
          "is_active",
          "usage_count",
          "rate_limit_per_hour",
          "rate_limit_per_day",
          "key"
        ],
        "title": "ApiKeyWithSecret",
        "description": "Schema for API key response with secret (only shown once)"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "example": {
          "detail": [
            {
              "loc": [
                "body",
                "email"
              ],
              "msg": "value is not a valid email address",
              "type": "value_error.email"
            }
          ]
        }
      },
...
      "PaginatedResponse_ApiKeyListResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiKeyListResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse[ApiKeyListResponse]"
      },
...
      "PaginatedResponse_UserListResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserListResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse[UserListResponse]"
      },
...
      "PaginatedResponse_WebhookListResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WebhookListResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit",
          "has_more"
        ],
        "title": "PaginatedResponse[WebhookListResponse]"
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "super_admin",
          "admin",
          "user",
          "viewer"
        ],
        "title": "UserRole",
        "description": "User role enumeration"
      },
      "UserStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive",
          "suspended",
          "deleted"
        ],
        "title": "UserStatus",
        "description": "User status enumeration"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "\n            JWT Authentication\n\n            To authenticate:\n            1. Login via `/api/v1/auth/login` to get an access token\n            2. Include the token in the Authorization header: `Bearer <token>`\n            3. Refresh the token using `/api/v1/auth/refresh` before it expires\n\n            Access tokens expire in 30 minutes.\n            Refresh tokens expire in 7 days.\n            "
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "\n            API Key Authentication\n\n            For programmatic access without user login:\n            1. Create an API key via `/api/v1/api-keys/`\n            2. Include the key in requests: `X-API-Key: <your_api_key>`\n\n            API keys can have custom scopes and IP restrictions.\n            "
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
...
}